+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/filter/IbzFavoritesSearchContext.java
@@ -34,8 +34,8 @@ public class IbzFavoritesSearchContext extends QueryWrapperContext<IbzFavorites>
             this.getSearchCond().like("ibz_favoritesname", n_ibz_favoritesname_like);
         }
     }
-	private Integer n_objectid_eq;//[数据对象标识]
-	public void setN_objectid_eq(Integer n_objectid_eq) {
+	private Long n_objectid_eq;//[数据对象标识]
+	public void setN_objectid_eq(Long n_objectid_eq) {
         this.n_objectid_eq = n_objectid_eq;
         if(!ObjectUtils.isEmpty(this.n_objectid_eq)){
             this.getSearchCond().eq("objectid", n_objectid_eq);
